Function to build funnel limits

Internal function for funnel plot to build the control limits prior to plotting.

build_limits_lookup(
  min_x,
  max_x,
  min_y,
  max_y,
  Poisson_limits,
  OD_adjust,
  tau2,
  data_type,
  sr_method,
  target,
  multiplier
)

Arguments

Poisson_limits
TRUE/FALSE Draw Poisson distribution limits?

OD_adjust
TRUE/FALSE Use overdispersion adjustment

tau2
If using OD_adjust, what is the tau2 ("between" standard error) to use?

data_type
SR, PR or RC. Used to set target reference

sr_method
Which adjustment method is being used, SHMI or CQC?

target
target to be used to set centre line

multiplier
Multiply ratio value by an amount. Default is 1, but some mortality ratios use 100, for example.

min_preds
Minimum predicted value for range of x-axis

max_preds
Maximum predicted value for range of x-axis

min_ratio
Minimum ratio value for range of y-axis

max_ratio
Maximum ratio value for range of y-axis

Value

A data.frame with an index column and various control limits based on the index as an x-axis value.