Overdispersion-adjusted funnel limit calculation

Add 95

calculate_limits(
  dfCI = dfCI,
  data_type = "SR",
  sr_method = "SHMI",
  multiplier = 1,
  tau2 = 0,
  target = target,
  draw_adjusted = draw_adjusted
)

Arguments

dfCI

Aggregated model input data

data_type

Type of data for adjustment and plotting: Indirectly Standardised ratio (\"SR\"), proportion (\"PR\"), or ratio of counts (\"RC\").

sr_method

Adjustment method for standardised ratios, can take the value \"SHMI\" or \"CQC\". \"SHMI\" is default.

multiplier

Multiplier to adjust limits if reporting by a multiplier, e.g. per 1000.

tau2

A 'between' standard deviation to add to the within standard deviation, S, to inflate limits.

target

The centre line of the plot. Mean for non-SRs or 1 for SR

Value

A data.frame of with appended OD limits